Mergebot functionality commit 1 - project structure and mergebot.py first shot.
