+++ config/apache.yaml
@@ -1,0 +1,6 @@
+--- 
+default_branch: master
+name: apache
+repository: apache
+scm_type: github
+verification_command: "mvn clean verify"
