+++ config/beam.yaml
@@ -1,0 +1,6 @@
+--- 
+default_branch: master
+name: apache_beam
+repository: incubator-beam
+scm_type: github
+verification_command: "mvn clean verify"
